requirements.txt
# Core
httpx>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# PDF processing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Token counting
tiktoken>=0.5.0

# API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
